{"schema":"decodex/commit/1","summary":"Extract frozen capture models","authority":"manual"}
